feature/email_rebuild

+++ persistence/src/main/java/pl/funnyqrz/repositories/ReportRepository.java
@@ -3,5 +3,7 @@ package pl.funnyqrz.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.funnyqrz.entities.ReportEntity;
 
-public interface ReportRepository extends JpaRepository<ReportEntity,Long> {
+import java.math.BigInteger;
+
+public interface ReportRepository extends JpaRepository<ReportEntity,BigInteger> {
 }
